% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_soil_table}
\alias{check_soil_table}
\title{Performs sanity checks on soil table}
\usage{
check_soil_table(soil_dt, nutrient)
}
\arguments{
\item{soil_dt}{A table where soil physical and chemical features needed are given or \code{NULL}}

\item{nutrient}{One of "nitrogen", "phosphorus", "potassium"}
}
\value{
A subset of the soil table featuring only the specific
analyses results for the \code{nutrient} given or, if \code{soil_dt} is \code{NULL}
a character vector of classes of the relevant soil features for each specific
nutrient.
}
\description{
The computation of nutrient fertilization plans requires
a few soil analyses be performed. Each nutrient fertilization plan
requires specific analyses, the functions checks whether these analyses
results are present in soil table for each specific nutrient.
}
\details{
The data object are passed to appropriate \code{fertplan} functions by \code{\link[=demand_nutrient]{demand_nutrient()}},
further information on soil features, and their unit of measure can be found in
help pages and vignettes of the \code{fertplan} package.
Templates for soil tables:

Template for \strong{nitrogen} nutrient table:
\itemize{
\item \strong{N_pc}, Percentage of nitrogen in soil, \code{numeric}
\item \strong{CNR}, Soil carbon / nitrogen ratio, \code{numeric}
\item \strong{SOM_pc}, Soil Organic Matter percentage, \code{numeric}
\item \strong{Clay_pc}, Soil clay content, \code{numeric}
}

Template for \strong{phosphorus} nutrient table:
\itemize{
\item \strong{P_ppm}, Phosphorus in soil in ppm (mg/kg), \code{numeric}
\item \strong{Limestone_pc}, Calcium in soil in percentage, \code{numeric}
}

templates for \strong{Potassium} nutrient table:
\itemize{
\item \strong{K_ppm}, Potassium in soil in ppm (mg/kg), \code{numeric}
\item \strong{Clay_pc}, Soil clay content, \code{numeric}
}

Note that the appropriate classes of soil features
for each specific nutrient may be
looked up by passing \code{NULL} as \code{soil_dt} argument, like in
\code{check_soil_table(NULL, "potassium")}.
}
\examples{
check_soil_table(NULL, "potassium")

}
