% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_vars}
\alias{check_vars}
\title{Performs sanity checks on soil variables}
\usage{
check_vars(vars = NULL, nutrient)
}
\arguments{
\item{vars}{A list where atmospheric, crop, and soil variables needed are given, or \code{NULL}}

\item{nutrient}{One of "nitrogen", "phosphorus", "potassium"}
}
\value{
A \code{data.table} subset of the \code{vars} list featuring only the specific
variables for the \code{nutrient} given or, if \code{vars} is \code{NULL}
a character vector of classes of the relevant variables for each specific
nutrient.
}
\description{
The computation of nutrient fertilization plans requires
a few atmospheric, crop, and soil variables to be shared among all soil analyses.
Each nutrient fertilization plan
requires specific variables, the functions checks whether these features
have been filled in the appropriate variable list, for each specific nutrient.
}
\details{
The data object are passed to appropriate \code{fertplan} functions by \code{\link[=demand_nutrient]{demand_nutrient()}},
further information on variables, and their unit of measure can be found in
help pages and vignettes of the \code{fertplan} package.
Variable list templates for variables:

Template for \strong{nitrogen} nutrient list:
\itemize{
\item \strong{crop}, The crop name to be sown, \code{character}
\item \strong{part}, The part of interest of the crop (Frutti, Pianta, ...), \code{character}
\item \strong{crop_type}, The crop type to be sown, \code{character}
\item \strong{expected_yield_kg_ha}, Expected crop yield, \code{integer}
\item \strong{prev_crop}, Soil clay content, \code{character}
\item \strong{texture}, Soil texture (one of "Sandy", "Loam", "Clayey"), \code{character}
\item \strong{drainage_rate}, Rate of drainage in soil (either "fast", "normal", "slow", "no drainage"), \code{character}
\item \strong{oct_jan_pr_mm}, cumulative precipitation in mm in the 4 months-period October - January, \code{integer}
\item \strong{n_supply_prev_frt_kg_ha}, Supply from organic fertilizations, \code{integer}
\item \strong{organic_fertilizer}, Type of organic fertilizer used, \code{character}
\item \strong{years_ago} * Time since last organic fertilization, \code{integer}
\item \strong{n_supply_atm_coeff}, A ratio to correct the N from atmosphere, \code{numeric}
}

Template for \strong{phosphorus} nutrient list:
\itemize{
\item \strong{crop}, The crop name to be sown, \code{character}
\item \strong{part}, The part of interest of the crop (Frutti, Pianta, ...), \code{character}
\item \strong{crop_class}, The class of crop to be sown, \code{character}
\item \strong{expected_yield_kg_ha}, Expected crop yield, \code{integer}
\item \strong{texture}, Soil texture (one of "Sandy", "Loam", "Clayey"), \code{character}
\item \strong{soil_depth_cm}, depth of soil tillage practise, \code{integer}
}

Template for \strong{potassium} nutrient list:
\itemize{
\item \strong{crop}, The crop name to be sown, \code{character}
\item \strong{part}, The part of interest of the crop (Frutti, Pianta, ...), \code{character}
\item \strong{crop_class}, The class of crop to be sown, \code{character}
\item \strong{expected_yield_kg_ha}, Expected crop yield, \code{integer}
\item \strong{texture}, Soil texture (one of "Sandy", "Loam", "Clayey"), \code{character}
\item \strong{soil_depth_cm}, depth of soil tillage practise, \code{integer}
}

Note that the appropriate classes of the variables
needed for each specific nutrient may be
looked up by passing \code{NULL} as \code{vars} argument, like in
\code{check_vars(NULL, "potassium")}.
}
\examples{
check_vars(NULL, "potassium")

}
