% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_nutrient.R
\name{demand_nutrient}
\alias{demand_nutrient}
\title{Title}
\usage{
demand_nutrient(soil_dt, vars, nutrient = "all")
}
\arguments{
\item{soil_dt}{a \code{data.table} of soil properties (columns) x soil samples (rows). Its features (soil analyses) depend on the nutrient fertilization specificed in \code{nutrient} parameter}

\item{vars}{a \code{list} of variables shared between soil samples. The specific analyses depend on the nutrient fertilization specified in \code{nutrient} parameter}

\item{nutrient}{a character vector detailing which nutrient fertilization plan to compute among "nitrogen", "phosphorus", and "potassium". Any combination of the three nutrients can be given or "All" (default) to computeall of them.}
}
\value{
a \code{data.table} with as many features as nutrients given as \code{nutrient} parameter and as many rows as soil properties given as \code{soil_dt} parameter.
}
\description{
More informations on soil properties and other variables related to the crop and the environment can be found in \code{\link{templates_l}}, in package \code{fertplan} help pages and vignettes and in the guideline of Region Lazio.
}
\seealso{
\code{fertplan} package.

Assessorato Agricoltura, Promozione della Filiera e della Cultura del Cibo, Ambiente e Risorse Naturali. 2020. “Parte Agronomica, Norme Generali, Disciplinare Di Produzione Integrata Della Regione Lazio - SQNPI.” Regione Lazio. \url{http://www.regione.lazio.it/rl_agricoltura/?vw=documentazioneDettaglio&id=52065}.
}
