% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertplanspatial-package.R
\docType{package}
\name{fertplanspatial}
\alias{fertplanspatial}
\title{fertplanspatial package}
\description{
The goal of the package is to define fertilization plans for the fields
of a farm and spatialize them. Fertilization plans rely on R package
`fertplan` to perform the raw computations.
}
\details{
It strictly follows the regulation for integrated crop production
published by the administrative authority of Lazio region, Italy.
Assessorato Agricoltura, Promozione della Filiera e della Cultura del
Cibo, Ambiente e Risorse Naturali, 2020. Parte agronomica, norme generali,
Disciplinare di Produzione Integrata della Regione Lazio - SQNPI.
Regione Lazio, Roma, Italy.
}
\seealso{
\url{http://www.regione.lazio.it/rl_agricoltura/?vw=documentazioneDettaglio&id=52065}

\code{fertplan} package
}
\author{
Marco Bascietto <marco.bascietto@crea.gov.it>
}
