% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soils}
\alias{soils}
\title{Soil physical and chemical analyses}
\format{
A soil samples \code{data.table} with 20 analyses (rows) and 13 features(columns):
\itemize{
\item \strong{id}: a simple id for each soil sample
\item \strong{X}: longitude coordinate in WGS84 Pseudo-Mercator Coordinate Reference System
\item \strong{Y}: latitude coordinate in WGS84 Pseudo-Mercator Coordinate Reference System
\item \strong{N_pc}: nitrogen content in \%
\item \strong{C_pc}: carbon content in \%
\item \strong{CNR}: carbon / nitrogen ratio
\item \strong{SOM_pc}: soil organic matter in \%
\item \strong{P_ppm}: phosphorus content in ppm (mg/kg)
\item \strong{K_ppm}: potassium content in ppm (mg/kg)
\item \strong{Limestone_pc}: calcium (Ca) content in \%
\item \strong{Clay_pc}: clay content in \%
}
}
\source{
\url{https://www.journals.elsevier.com/european-journal-of-agronomy}
}
\usage{
soils
}
\description{
A dataset containing the results of physical and chemical analyses on 20
samples of soil in a field located in the property of the Italian Council
for Agricultural Research and Economica, Research Centre for Engineering
and Agro-Food Processing of Monterotondo (Rome).
}
\details{
The soil sample points are randomly spread over a 5 ha field previously cultivated on
wheat followed by a few month-long fallow. Each sample point is spatially located
in WGS84 Pseudo-Mercator coordinates (unit: metres) and uniquely identified
through a numeric \code{id}. Soil samples were analyzed in October 2019.
}
\keyword{datasets}
