% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soils_spatial}
\alias{soils_spatial}
\title{Spatial soil nutrients fertilization dataset}
\format{
A soil sample points \code{data.table} with 20 sample points (rows)
and 6 features(columns):
\itemize{
\item \strong{id}: a simple id for each soil sample point
\item \strong{X}: longitude coordinate in WGS84 Pseudo-Mercator Coordinate Reference System
\item \strong{Y}: latitude coordinate in WGS84 Pseudo-Mercator Coordinate Reference System
\item \strong{nitrogen}: estimate of nitrogen demand in kg/ha
\item \strong{phosphorus}: estimate of phosphorus (P_2O) demand in kg/ha
\item \strong{potassium}: estimate of potassium (K_2O_5) demand in kg/ha
}
}
\source{
\url{https://www.journals.elsevier.com/european-journal-of-agronomy}
}
\usage{
soils_spatial
}
\description{
A \code{\link[sp:SpatialPointsDataFrame]{sp::SpatialPointsDataFrame()}} dataset containing the results
of applying \code{\link[=demand_nutrient]{demand_nutrient()}} on \code{\link{soils}} dataset to get
NPK fertilization plans.
}
\details{
Each soil sample point is spatially located in WGS84 Pseudo-Mercator coordinates
(epsg:3857, unit: metres) and uniquely identified through a numeric \code{id}.
}
\keyword{datasets}
