% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_nutrient.R
\name{spatial_nutrient}
\alias{spatial_nutrient}
\title{Spatialize nutrient fertilization plans}
\usage{
spatial_nutrient(
  sp_df,
  model = "auto",
  grid_spdf = NULL,
  bbox_buffer = 0L,
  spat_res = 5L,
  nutrient = "all",
  ...
)
}
\arguments{
\item{sp_df}{a Spatial soil nutrients fertilization dataset}

\item{model}{a character vector describing the variogram model type.
Either "auto" or one of the available variogram models to \code{\link[gstat:vgm]{gstat::vgm()}},
e.g. "Exp", "Sph", "Gau", "Mat". Specific model parameters may be passed
to \code{\link[gstat:vgm]{gstat::vgm()}} as named function arguments arguments.
In "auto" the optimal variogram
is chosen by \code{\link[automap:autofitVariogram]{automap::autofitVariogram()}}. Calling gstat::vgm() without a model
argument returns a data.frame with available models.}

\item{grid_spdf}{a \code{\link[sp:SpatialPoints]{sp::SpatialPoints()}} object grid with nutrient predictions
locations. If \code{NULL} (default) then a point grid is built using \code{sp_df} bounding
box and coordinate reference system, each point spatially distant by \code{spat_res}.}

\item{bbox_buffer}{an integer value to enlarge (or shrink) the bounding box the
default point grid is built in (i.e. when grid_spdf is set to \code{NULL}). Default to \code{0L},
i.e. no effect on the bounding box.}

\item{spat_res}{a positive number indicating the spatial resolution
of the resulting spatial grid when \code{grid_spdf} is \code{NULL}. Default is 5 metres.}

\item{nutrient}{a character vector detailing which nutrient
spatialized fertilization plan to compute among "nitrogen",
"phosphorus", and "potassium". Any combination of the three
nutrients can be given or "all" (default) to compute all of them.}

\item{...}{arguments that will be passed to \code{\link[gstat:vgm]{gstat::vgm()}}}
}
\value{
a list with as many elements as nutrients given as \code{nutrient}
parameter. Each element is a \code{\link[sp:SpatialPointsDataFrame]{sp::SpatialPointsDataFrame()}}.
}
\description{
Spatialization is performed bny fitting a variogram model to the nutrients
fertilization plans and then performing ordinary kriging.
}
\details{
The function returns a list of \code{\link[sp:SpatialPointsDataFrame]{sp::SpatialPointsDataFrame()}} grids on the
bounding box of the spatial soil nutrients fertilization dataset, as
returned by \code{\link[=demand_nutrient]{demand_nutrient()}}. Each spatial grid can be readily plotted or
rasterized.

Spatial coordinates must be passed as 'X' and 'Y' named columns in \code{sp_df}
spatial table. This table should also containing either 'nitrogen',
'phosphorus', and 'potassium' named columns according to the nutrient plan
to be spatialized passed as \code{nutrient} argument.
}
