% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_nutrient.R
\name{to_sp}
\alias{to_sp}
\title{Tranform a data.frame into a SpatialPoints}
\usage{
to_sp(tab, crs_s)
}
\arguments{
\item{tab}{a data.fram typically build via \code{\link[=expand.grid]{expand.grid()}}}

\item{crs_s}{a proj4string coordinate reference system typically returned by \code{\link[sp:proj4string]{sp::proj4string()}}}
}
\value{
a \code{\link[sp:SpatialPoints]{sp::SpatialPoints()}} for \code{tab} with a crs assigned and X, Y coordinates identified
}
\description{
Assign a Coordinate Reference System proj4 string and identify geographical
coordinates in a \code{data.frame} to return a SpatialPoints object. Geographical coordinates
must be present in the \code{data.frame} and named \code{X}, and \code{Y}.
}
